refactor(api): move checkCampaignParticipationBelongsToUser from shared to prescription
